# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Digitalisasi Arsip",
        "description": "REST API untuk sistem digitalisasi arsip akademik.\n\n## Autentikasi\nAPI ini menggunakan Laravel Sanctum dengan session-based authentication / Bearer Token.\n\n**Langkah autentikasi (Session):**\n1. GET `/api/csrf-cookie` untuk mendapatkan CSRF token\n2. POST `/api/auth/login` dengan credentials\n3. Sertakan cookie dan X-XSRF-TOKEN header pada setiap request\n\n## Roles\n- **Manager**: Full access ke semua fitur\n- **Uploader**: Upload, update & delete dokumen sendiri\n- **QC**: Verifikasi dokumen\n- **SBAP**: Download dokumen terverifikasi",
        "contact": {
            "name": "API Support"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/reports/generate": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate Report (UC-09)",
                "description": "Membuat laporan berdasarkan parameter yang diberikan (Manager only).\n\nMendukung format PDF, XLSX, dan CSV.\nTipe laporan: monthly, annual, custom.",
                "operationId": "generateReport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportGenerateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File laporan berhasil di-generate",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal membuat laporan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal membuat laporan. Silakan coba lagi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/reports/dashboard": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Dashboard Stats (UC-09)",
                "description": "Mengambil statistik dashboard untuk halaman laporan (Manager only)",
                "operationId": "getDashboardStats",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter dari tanggal (default: awal bulan ini)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter sampai tanggal (default: sekarang)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistik dashboard berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardStatsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/reports/qc-performance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan performa dan efisiensi staf verifikator QC (UC-16) - Manager only",
                "description": "Mengambil ringkasan statistik dan detail kinerja individual staf QC (jumlah verifikasi, penolakan, akurasi, status online/last seen) dengan paginasi dan filter periode. Hanya dapat diakses oleh Manager.",
                "operationId": "6e176897015e76b20f70f638dfeec2b4",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Tanggal awal periode (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Tanggal akhir periode (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-18"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah verifikator per halaman (default 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan performa QC berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QcPerformanceReportResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (Bukan Manager)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationError": {
                "properties": {
                    "message": {
                        "description": "Reusable OpenAPI Schema Definitions.\n\nFile ini berisi semua schema yang digunakan secara berulang\ndi berbagai endpoint (request bodies, responses, dll).",
                        "type": "string",
                        "example": "The email field is invalid."
                    },
                    "errors": {
                        "description": "Object berisi field yang error dengan array pesan error",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 72
                    }
                },
                "type": "object"
            },
...
            "ReportGenerateRequest": {
                "required": [
                    "period_start",
                    "period_end",
                    "format",
                    "type"
                ],
                "properties": {
                    "period_start": {
                        "description": "Tanggal awal periode laporan",
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01"
                    },
                    "period_end": {
                        "description": "Tanggal akhir periode laporan",
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-31"
                    },
                    "format": {
                        "description": "Format file laporan",
                        "type": "string",
                        "example": "pdf",
                        "enum": [
                            "pdf",
                            "xlsx",
                            "csv"
                        ]
                    },
                    "type": {
                        "description": "Tipe laporan",
                        "type": "string",
                        "example": "monthly",
                        "enum": [
                            "monthly",
                            "annual",
                            "custom"
                        ]
                    },
                    "style": {
                        "description": "Gaya laporan (opsional)",
                        "type": "string",
                        "example": "detailed",
                        "nullable": true,
                        "enum": [
                            "detailed",
                            "summary",
                            "executive"
                        ]
                    },
                    "content": {
                        "description": "Konten yang dimasukkan ke laporan (opsional)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "upload_stats",
                                "qc_metrics",
                                "doc_status",
                                "user_activity",
                                "trend_analysis"
                            ]
                        },
                        "example": [
                            "upload_stats",
                            "qc_metrics",
                            "doc_status"
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DashboardStatsResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Statistik dashboard berhasil diambil."
                    },
                    "data": {
                        "properties": {
                            "total_reports_generated": {
                                "type": "integer",
                                "example": 0
                            },
                            "most_downloaded_type": {
                                "type": "string",
                                "example": "Bulanan"
                            },
                            "last_generated": {
                                "type": "string",
                                "example": "2026-02-22 00:20"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "QcPerformanceItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "QC Verifier 1"
                    },
                    "email": {
                        "type": "string",
                        "example": "qc1@test.com"
                    },
                    "is_online": {
                        "type": "boolean",
                        "example": true
                    },
                    "last_seen_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-18T20:00:00.000000Z",
                        "nullable": true
                    },
                    "total_processed": {
                        "type": "integer",
                        "example": 50
                    },
                    "verified_count": {
                        "type": "integer",
                        "example": 45
                    },
                    "rejected_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "accuracy_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 90
                    }
                },
                "type": "object"
            },
            "QcPerformanceSummary": {
                "properties": {
                    "total_qc_staff": {
                        "type": "integer",
                        "example": 4
                    },
                    "active_online_qc": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_verified": {
                        "type": "integer",
                        "example": 120
                    },
                    "total_rejected": {
                        "type": "integer",
                        "example": 15
                    },
                    "average_accuracy_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 88.89
                    }
                },
                "type": "object"
            },
            "QcPerformanceReportResponse": {
                "properties": {
                    "summary": {
                        "$ref": "#/components/schemas/QcPerformanceSummary"
                    },
                    "verifiers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QcPerformanceItem"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "period": {
                        "properties": {
                            "start_date": {
                                "type": "string",
                                "example": "2026-01-01"
                            },
                            "end_date": {
                                "type": "string",
                                "example": "2026-08-18"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "Session cookie dari Laravel Sanctum",
                "name": "digitalisasi_arsip_api_session",
                "in": "cookie"
            }
        }
    },
...
}